+++ app/Http/Controllers/FileController.php
@@ -4,7 +4,7 @@ namespace Servidor\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use InvalidArgumentException;
+use Illuminate\Validation\ValidationException;
 use Servidor\FileManager\FileManager;
 
 class FileController extends Controller
@@ -33,28 +33,21 @@ class FileController extends Controller
                 return response($file, $file['error']['code']);
             }
 
-            try {
-                return response()->json($file);
-            } catch (InvalidArgumentException $e) {
-                $file['contents'] = '';
-                $file['error'] = [
-                    'code' => 422,
-                    'msg' => 'Failed loading file',
-                ];
-
-                return response()->json($file);
-            }
+            return response()->json($file);
         }
 
         $path = $request->get('path');
+        $list = $this->fm->list($path);
 
-        return response()->json($this->fm->list($path));
+        return isset($list['error'])
+            ? response($list, $list['error']['code'])
+            : response()->json($list);
     }
 
     public function update(Request $request)
     {
         if (!$filepath = $request->query('file')) {
-            throw new InvalidArgumentException('File path must be specified.');
+            throw ValidationException::withMessages(['file' => 'File path must be specified.']);
         }
 
         $data = $request->validate([