+++ resources/js/components/Files/Browser/FileList.vue
@@ -1,5 +1,5 @@
 <template>
-    <sui-table compact selectable :inverted="darkMode" class="files">
+    <sui-table compact selectable :inverted="darkMode" class="files" v-if="!files.error">
         <thead>
             <tr>
                 <th>Filename</th>
@@ -13,6 +13,13 @@
                 :file="file" :path="path" />
         </tbody>
     </sui-table>
+    <sui-segment class="placeholder" :inverted="darkMode" v-else>
+        <sui-header icon>
+            <sui-icon v-if="files.error.code == 404" name="search" color="teal" />
+            <sui-icon v-else name="bug" color="orange" />
+            {{ files.error.msg }}
+        </sui-header>
+    </sui-segment>
 </template>
 
 <script>