+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Deleting system users and groups now triggers a confirmation prompt
 * Home directory can now be purged when deleting a system user
 
+### Fixed
+* File contents no longer have whitespace and newlines trimmed when saved
+* Listing a directory that doesn't exist will now show a friendly error
+
 
 ## [0.8.0] - 2020-02-02
 ### Added